+++ custom/Gaffer/Gaffer.py
@@ -91,6 +91,9 @@ class GafferPlugin(DeadlinePlugin):
         # Generic Gaffer progress and error
         self.AddStdoutHandlerCallback(".*Progress: (\d+)%.*").HandleCallback += self.HandleProgress
 
+        # Alfred style progress from Houdini
+        self.AddStdoutHandlerCallback(".*ALF_PROGRESS\s*(\d+)%.*").HandleCallback += self.HandleProgress
+
         # Vray's ply2vrmesh prints out lines for each frame and also each voxel within the frame
         self.AddStdoutHandlerCallback(".*Subdividing frame ([0-9]+) of ([0-9]+).*").HandleCallback += self.HandlePly2VrmeshFrameProgress
         self.AddStdoutHandlerCallback(".*Processing voxel ([0-9]+) of ([0-9]+).*").HandleCallback += self.HandlePly2VrmeshVoxelProgress